@winder 2021-03-23
Thanks, Fabrice. I suppose we should add this for the AssetHolding, Application Global State and Application Local State tables also.